+++ xbmc/addons/AddonStatusHandler.cpp
@@ -18,11 +18,11 @@
  *
  */
 #include "AddonStatusHandler.h"
-#include "AddonManager.h"
+#include "addons/AddonManager.h"
+#include "addons/settings/GUIDialogAddonSettings.h"
 #include "threads/SingleLock.h"
 #include "messaging/ApplicationMessenger.h"
 #include "guilib/GUIWindowManager.h"
-#include "GUIDialogAddonSettings.h"
 #include "dialogs/GUIDialogYesNo.h"
 #include "dialogs/GUIDialogOK.h"
 #include "dialogs/GUIDialogKaiToast.h"
