+++ xbmc/addons/CMakeLists.txt
@@ -16,7 +16,6 @@ set(SOURCES Addon.cpp
             FilesystemInstaller.cpp
             GameResource.cpp
             GUIDialogAddonInfo.cpp
-            GUIDialogAddonSettings.cpp
             GUIViewStateAddonBrowser.cpp
             GUIWindowAddonBrowser.cpp
             ImageDecoder.cpp
@@ -57,7 +56,6 @@ set(HEADERS Addon.h
             FilesystemInstaller.h
             GameResource.h
             GUIDialogAddonInfo.h
-            GUIDialogAddonSettings.h
             GUIViewStateAddonBrowser.h
             GUIWindowAddonBrowser.h
             IAddon.h
