+++ xbmc/addons/settings/CMakeLists.txt
@@ -1,7 +1,9 @@
 set(SOURCES AddonSettings.cpp
+            GUIDialogAddonSettings.cpp
             SettingUrlEncodedString.cpp)
 
 set(HEADERS AddonSettings.h
+            GUIDialogAddonSettings.h
             SettingUrlEncodedString.h)
 
 core_add_library(addonsSettings)
