+++ xbmc/addons/settings/GUIDialogAddonSettings.cpp
