+++ xbmc/addons/settings/GUIDialogAddonSettings.h
