+++ xbmc/settings/dialogs/GUIDialogContentSettings.cpp
@@ -28,7 +28,7 @@
 
 #include "GUIDialogContentSettings.h"
 #include "addons/AddonSystemSettings.h"
-#include "addons/GUIDialogAddonSettings.h"
+#include "addons/settings/GUIDialogAddonSettings.h"
 #include "addons/GUIWindowAddonBrowser.h"
 #include "filesystem/AddonsDirectory.h"
 #include "dialogs/GUIDialogKaiToast.h"
